re-enable ncclep tests and consolidate the gates
