+++ tests/unit_tests/a2a_overlap/test_fsdp_1f1b_overlap.py
@@ -26,9 +26,6 @@ from tests.unit_tests.a2a_overlap.utils import (
 )
 from tests.unit_tests.test_utilities import Utils
 
-# Transformer Engine 2.17 aborts in the A2A overlap suite with a pybind11 GIL dec_ref failure.
-pytestmark = pytest.mark.flaky_in_dev
-
 SEQ_LEN = 32
 VOCAB_SIZE = 128
 NUM_STEPS = 3
