+++ tests/unit_tests/a2a_overlap/test_schedule_chunk_1f1b.py
@@ -23,9 +23,6 @@ from tests.unit_tests.a2a_overlap.utils import (
 )
 from tests.unit_tests.test_utilities import Utils
 
-# Transformer Engine 2.17 aborts in the A2A overlap suite with a pybind11 GIL dec_ref failure.
-pytestmark = pytest.mark.flaky_in_dev
-
 
 def build_model(config, use_padding_mask=False):
     seq_len = 32
