+++ tests/unit_tests/transformer/test_cuda_graphs.py
@@ -66,7 +66,11 @@ from megatron.training.global_vars import (
     set_global_variables,
 )
 from megatron.training.training import setup_model_and_optimizer
-from tests.unit_tests.test_utilities import Utils
+from tests.unit_tests.test_utilities import (
+    Utils,
+    is_nccl_ep_available,
+    is_nccl_ep_fp8_dispatch_available,
+)
 
 fp8_available, _ = check_fp8_support()
 
@@ -1406,12 +1410,6 @@ def is_hybrid_ep_available():
     return HAVE_HYBRIDEP
 
 
-def is_nccl_ep_available():
-    from megatron.core.transformer.moe.fused_a2a import HAVE_TE_EP
-
-    return HAVE_TE_EP
-
-
 class TestPartialCudaGraph:
     """Test that CUDA graph outputs match non-CUDA graph outputs for various scopes."""
 
@@ -1673,10 +1671,6 @@ class TestPartialCudaGraph:
             extra_kwargs["moe_token_dispatcher_type"] = "flex"
             extra_kwargs["moe_flex_dispatcher_backend"] = "ncclep"
         elif moe_dispatcher_type == "ncclep_fp8":
-            from tests.unit_tests.transformer.moe.test_token_dispatcher import (
-                is_nccl_ep_fp8_dispatch_available,
-            )
-
             if not is_nccl_ep_available():
                 pytest.skip("NCCL EP is not available")
             if ep_size < 2:
